Moves options, records and statistics UI to their own folders.
