+++ src/gui/MineSwept.java
@@ -3,9 +3,9 @@ package gui;
 import gui.events.EventPublisher;
 import gui.events.QuitGameEvent;
 import gui.menu.Menus;
-import gui.menu.OptionWindow;
-import gui.menu.RecordWindow;
-import gui.menu.StatisticsWindow;
+import gui.options.OptionWindow;
+import gui.records.RecordWindow;
+import gui.statistics.StatisticsWindow;
 import gui.panel.MainPanel;
 import java.awt.*;
 import java.awt.event.WindowEvent;
