+++ src/gui/events/ShowStatisticsEvent.java
@@ -1,7 +1,5 @@
 package gui.events;
 
-import gui.menu.StatisticsWindow;
-
 public class ShowStatisticsEvent {
 	public boolean show;
