+++ src/gui/records/RecordPanel.java
@@ -1,4 +1,4 @@
-package gui.menu;
+package gui.records;
 
 import logic.files.Records;
 import javax.swing.*;
